backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
httpx
google-generativeai
Pillow
